+++ src/storage/mod.rs
@@ -29,18 +29,23 @@ impl Storage {
         self.inner.id()
     }
 
-    async fn get(&self, key: &str) -> crate::Result<String> {
-        self.inner.get(key).await.and_then(|record| {
-            if let Some(key) = &self.encryption_key {
-                if serde_json::from_str::<Vec<u8>>(&record).is_ok() {
-                    Ok(String::from_utf8_lossy(&chacha::aead_decrypt(key, record.as_bytes())?).into_owned())
+    async fn get(&self, key: &str) -> crate::Result<Option<String>> {
+        match self.inner.get(key).await? {
+            Some(record) => {
+                if let Some(key) = &self.encryption_key {
+                    if serde_json::from_str::<Vec<u8>>(&record).is_ok() {
+                        Ok(Some(
+                            String::from_utf8_lossy(&chacha::aead_decrypt(key, record.as_bytes())?).into_owned(),
+                        ))
+                    } else {
+                        Ok(Some(record))
+                    }
                 } else {
-                    Ok(record)
+                    Ok(Some(record))
                 }
-            } else {
-                Ok(record)
             }
-        })
+            None => Ok(None),
+        }
     }
 
     async fn set<T: Serialize + Send>(&mut self, key: &str, record: T) -> crate::Result<()> {