+++ CHANGELOG.md
@@ -24,6 +24,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Updated dependencies;
+- `StorageAdapter::get` return type changed to `Result<Option<String>>`;
+- `StorageManager::get_account_manager_data` return type changed to `Result<Option<AccountManagerBuilder>>`;
 
 ### Removed